requirement.txt
pandas==2.2.2
scikit_learn==1.5.0 
matplotlib==3.9.0
seaborn==0.13.2
xgboost-2.1.0
tqdm-4.66.4
optuna-3.6.1
flask-3.0.3
flask_wtf-1.2.1
nbimporter-0.3.4